Fix linter errors and warnings (#698)
Fixes or ignores errors and warnings (not refactors or conventions) reported by pylint
