+++ app/forms/duration_form.py
@@ -9,11 +9,13 @@ from app.forms.fields import IntegerFieldWithSeparator
 ErrorMessageType = dict[str, str]
 
 
+# pylint: disable=no-member
+# wtforms Form parents are not discoverable in the 2.3.3 implementation
 class DurationForm(Form):
     def validate(
         self, extra_validators: Optional[dict[str, list[Callable]]] = None
     ) -> bool:
-        super(DurationForm, self).validate(extra_validators)
+        super().validate(extra_validators)
 
         if all(not field.raw_data[0] for field in self._fields.values()):
             if self.mandatory:
