+++ app/forms/fields/month_year_date_field.py
@@ -13,6 +13,8 @@ def get_form_class(validators):
 
         @cached_property
         def data(self):
+            # pylint: disable=no-member
+            # wtforms Form parents are not discoverable in the 2.3.3 implementation
             data = super().data
 
             try:
