+++ app/forms/fields/year_date_field.py
@@ -12,6 +12,8 @@ def get_form_class(validators):
 
         @cached_property
         def data(self):
+            # pylint: disable=no-member
+            # wtforms Form parents are not discoverable in the 2.3.3 implementation
             data = super().data
 
             try:
