+++ app/routes/session.py
@@ -54,7 +54,7 @@ def login():
         runner_claims = validate_runner_claims(decrypted_token)
     except ValidationError as e:
         raise InvalidTokenException("Invalid runner claims") from e
-
+    # pylint: disable=assigning-non-slot
     g.schema = load_schema_from_metadata(runner_claims)
     schema_metadata = g.schema.json["metadata"]
