+++ app/views/handlers/feedback.py
@@ -75,6 +75,8 @@ class Feedback:
         return build_feedback_context(self.question_schema, self.form)
 
     def get_page_title(self) -> str:
+        # pylint: disable=no-member
+        # wtforms Form parents are not discoverable in the 2.3.3 implementation
         if self.form.errors:
             title: str = gettext("Error: {page_title}").format(
                 page_title=self.PAGE_TITLE
@@ -90,7 +92,8 @@ class Feedback:
             session_data.case_id,
             session_data.tx_id,
         )
-
+        # pylint: disable=no-member
+        # wtforms Form parents are not discoverable in the 2.3.3 implementation
         feedback_message = FeedbackPayload(
             metadata=self._questionnaire_store.metadata,
             response_metadata=self._questionnaire_store.response_metadata,
