+++ app/views/handlers/list_collector.py
@@ -47,6 +47,8 @@ class ListCollector(Question):
 
         if answer_action and answer_action["type"] == "RedirectToListAddBlock":
             self._is_adding = True
+            # pylint: disable=no-member
+            # wtforms Form parents are not discoverable in the 2.3.3 implementation
             self.questionnaire_store_updater.update_answers(self.form.data)
             self.questionnaire_store_updater.save()
         else:
