+++ app/views/handlers/question.py
@@ -131,6 +131,8 @@ class Question(BlockHandler):
         answers = self.rendered_block["question"]["answers"]
 
         for answer in answers:
+            # pylint: disable=no-member
+            # wtforms Form parents are not discoverable in the 2.3.3 implementation
             submitted_answer = self.form.data[answer["id"]]
 
             for option in answer.get("options", {}):
@@ -151,7 +153,8 @@ class Question(BlockHandler):
 
         if "list_summary" in self.rendered_block:
             context.update(self.get_list_summary_context())
-
+        # pylint: disable=no-member
+        # wtforms Form parents are not discoverable in the 2.3.3 implementation
         if self.form.errors or self.form.question_errors:
             self.page_title = gettext("Error: {page_title}").format(
                 page_title=self.page_title
@@ -178,6 +181,8 @@ class Question(BlockHandler):
         )
 
     def handle_post(self):
+        # pylint: disable=no-member
+        # wtforms Form parents are not discoverable in the 2.3.3 implementation
         self.questionnaire_store_updater.update_answers(self.form.data)
         if self.questionnaire_store_updater.is_dirty():
             self._routing_path = self.router.routing_path(
