+++ scripts/run_lint_python.sh
@@ -21,11 +21,11 @@ function display_result {
 flake8 --max-complexity 10 --count
 display_result $? 1 "Flake 8 code style check"
 
-pylint --reports=n --output-format=colorized --rcfile=.pylintrc -j 0 **/*.py *.py
 # pylint bit encodes the exit code to allow you to figure out which category has failed.
 # https://docs.pylint.org/en/1.6.0/run.html#exit-codes
 # We want to fail on all errors so don't check for specific bits in the output; but if we did in future, see:
 # http://stackoverflow.com/questions/6626351/how-to-extract-bits-from-return-code-number-in-bash
+find . -type f -name "*.py" | xargs pylint --reports=n --output-format=colorized --rcfile=.pylintrc -j 0 --disable=R,C
 display_result $? 1 "Pylint linting check"
 
 isort --check .
