+++ tests/app/cloud_tasks/test_cloud_task_publishers.py
@@ -10,6 +10,7 @@ from app.cloud_tasks import CloudTaskPublisher
 from app.cloud_tasks.exceptions import CloudTaskCreationFailed
 
 
+# pylint: disable=protected-access
 class TestCloudTaskPublisher(TestCase):
     PROJECT_ID = "test-project-id"
     queue_name = "test"
