+++ tests/app/forms/field_handlers/test_text_area_handler.py
@@ -35,10 +35,10 @@ def test_get_field(value_source_resolver):
 
 
 def test_get_length_validator(value_source_resolver):
-    error_messages = {
+    test_error_messages = {
         "MAX_LENGTH_EXCEEDED": "This is the default max length of %(max)d message"
     }
-    text_area_handler = TextAreaHandler({}, value_source_resolver, error_messages)
+    text_area_handler = TextAreaHandler({}, value_source_resolver, test_error_messages)
     validator = text_area_handler.get_length_validator()
 
     assert validator.message == "This is the default max length of %(max)d message"
@@ -52,8 +52,12 @@ def test_get_length_validator_with_message_override(value_source_resolver):
             }
         }
     }
-    error_messages = {"MAX_LENGTH_EXCEEDED": "This is the default max length message"}
-    text_area_handler = TextAreaHandler(answer, value_source_resolver, error_messages)
+    test_error_messages = {
+        "MAX_LENGTH_EXCEEDED": "This is the default max length message"
+    }
+    text_area_handler = TextAreaHandler(
+        answer, value_source_resolver, test_error_messages
+    )
 
     validator = text_area_handler.get_length_validator()
 
@@ -62,8 +66,10 @@ def test_get_length_validator_with_message_override(value_source_resolver):
 
 def test_get_length_validator_with_max_length_override(value_source_resolver):
     answer = {"max_length": 30}
-    error_messages = {"MAX_LENGTH_EXCEEDED": "%(max)d characters"}
-    text_area_handler = TextAreaHandler(answer, value_source_resolver, error_messages)
+    test_error_messages = {"MAX_LENGTH_EXCEEDED": "%(max)d characters"}
+    text_area_handler = TextAreaHandler(
+        answer, value_source_resolver, test_error_messages
+    )
     validator = text_area_handler.get_length_validator()
 
     assert validator.max == 30
