+++ tests/app/views/contexts/test_thank_you_context.py
@@ -1,9 +1,7 @@
 from datetime import datetime, timedelta, timezone
 
-import pytest
 from flask import Flask
 
-from app.data_models.session_data import SessionData
 from app.utilities.schema import load_schema_from_name
 from app.views.contexts.thank_you_context import build_thank_you_context
 
@@ -11,22 +9,6 @@ SURVEY_TYPE_DEFAULT = "default"
 SURVEY_TYPE_SOCIAL = "social"
 SUBMITTED_AT = datetime.now(timezone.utc)
 SCHEMA = load_schema_from_name("test_view_submitted_response", "en")
-
-
-@pytest.fixture
-def fake_session_data():
-    return SessionData(
-        tx_id="tx_id",
-        schema_name="some_schema_name",
-        language_code=None,
-        launch_language_code=None,
-        survey_url=None,
-        ru_ref="ru_ref",
-        response_id="response_id",
-        case_id="case_id",
-        period_str=None,
-        ru_name=None,
-    )
 
 
 def test_social_survey_context(fake_session_data, app: Flask):
