+++ tests/app/views/handlers/test_view_submitted_response_pdf.py
@@ -9,7 +9,9 @@ from app.views.handlers.view_submitted_response_pdf import ViewSubmittedResponse
 from .conftest import set_storage_data
 
 
-def test_pdf_not_downloadable(app, storage, schema, language):
+def test_pdf_not_downloadable(
+    app, storage, schema, language  # pylint:disable=unused-argument
+):
     submitted_at = datetime.now(timezone.utc) - timedelta(minutes=46)
     set_storage_data(storage, submitted_at=submitted_at)
 
@@ -23,7 +25,9 @@ def test_pdf_not_downloadable(app, storage, schema, language):
         )
 
 
-def test_filename_uses_schema_name(app, storage, schema, language):
+def test_filename_uses_schema_name(
+    app, storage, schema, language  # pylint:disable=unused-argument
+):
     submitted_at = datetime.now(timezone.utc)
     set_storage_data(storage, submitted_at=submitted_at)
