+++ tests/functional/generate_pages.py
@@ -223,13 +223,11 @@ LIST_SECTION_SUMMARY_ADD_LINK_GETTER = Template(
 
 """
 )
-
 LIST_SECTION_SUMMARY_EDIT_LINK_GETTER = Template(
     r"""  ${list_name}ListEditLink(listItemInstance) { return `div[data-qa="${list_name}-list-summary"] a[data-qa="list-item-change-` + listItemInstance + `-link"]`; }
 
 """
 )
-
 LIST_SECTION_SUMMARY_REMOVE_LINK_GETTER = Template(
     r"""  ${list_name}ListRemoveLink(listItemInstance) { return `div[data-qa="${list_name}-list-summary"] a[data-qa="list-item-remove-` + listItemInstance + `-link"]`; }
 
@@ -414,15 +412,13 @@ def process_calculated_summary(answers, page_spec):
         page_spec.write(CALCULATED_SUMMARY_LABEL_GETTER.substitute(answer_context))
 
 
-def process_final_summary(
-    schema_data, require_path, dir_out, spec_file, collapsible, section_summary=False
-):
+def process_final_summary(schema_data, require_path, dir_out, spec_file, collapsible):
     page_filename = "submit.page.js"
     page_path = os.path.join(dir_out, page_filename)
 
     logger.info("creating %s...", page_path)
 
-    with open(page_path, "w") as page_spec:
+    with open(page_path, "w", encoding="utf-8") as page_spec:
         block_context = build_and_get_base_page_context(
             page_dir=dir_out.split("/")[-1],
             page_spec=page_spec,
@@ -454,7 +450,7 @@ def process_view_submitted_response(schema_data, require_path, dir_out, spec_fil
 
     logger.info("creating %s...", page_path)
 
-    with open(page_path, "w") as page_spec:
+    with open(page_path, "w", encoding="utf-8") as page_spec:
         block_context = build_and_get_base_page_context(
             page_dir=dir_out.split("/")[-1],
             page_spec=page_spec,
@@ -700,7 +696,7 @@ def process_block(
 
     logger.info("creating %s...", page_path)
 
-    with open(page_path, "w") as page_spec:
+    with open(page_path, "w", encoding="utf-8") as page_spec:
         page_name = generate_pascal_case_from_id(block["id"])
 
         base_page = "QuestionPage"
@@ -776,8 +772,8 @@ def _has_definitions_in_block_contents(block_contents):
 
 def process_schema(in_schema, out_dir, spec_file, require_path=".."):
     try:
-        data = json_loads(open(in_schema).read())
-    except Exception as ex:
+        data = json_loads(open(in_schema, encoding="utf-8").read())
+    except Exception:  # pylint: disable=broad-except
         logger.error("error reading %s", in_schema)
         return
 
@@ -816,7 +812,6 @@ def process_questionnaire_flow(schema_data, require_path, dir_out, spec_file):
             dir_out,
             spec_file,
             collapsible,
-            section_summary=False,
         )
 
 
@@ -832,7 +827,7 @@ def process_section_summary(
 
     logger.info("creating %s...", page_path)
 
-    with open(page_path, "w") as page_spec:
+    with open(page_path, "w", encoding="utf-8") as page_spec:
 
         section_context = {
             "pageName": generate_pascal_case_from_id(section_id),
@@ -862,7 +857,7 @@ def process_section_summary(
 
 
 def append_spec_page_import(context, spec_file):
-    with open(spec_file, "a") as required_template_spec:
+    with open(spec_file, "a", encoding="utf-8") as required_template_spec:
         required_template_spec.write(SPEC_PAGE_IMPORT.substitute(context))
 
 
@@ -875,7 +870,7 @@ if __name__ == "__main__":
 
     if template_spec_file:
         os.makedirs(os.path.dirname(template_spec_file), exist_ok=True)
-        with open(template_spec_file, "w") as template_spec:
+        with open(template_spec_file, "w", encoding="utf-8") as template_spec:
             template_spec.write(SPEC_PAGE_HEADER)
             template_spec.close()
 
@@ -883,7 +878,7 @@ if __name__ == "__main__":
                 args.SCHEMA, args.OUT_DIRECTORY, template_spec_file, args.require_path
             )
 
-            with open(template_spec_file, "a") as template_spec:
+            with open(template_spec_file, "a", encoding="utf-8") as template_spec:
                 schema_name = {"schema": os.path.basename(args.SCHEMA)}
                 template_spec.write(SPEC_EXAMPLE_TEST.substitute(schema_name))
     else:
