+++ tests/integration/integration_test_case.py
@@ -27,7 +27,7 @@ KEYS_FOLDER = "./tests/jwt-test-keys"
 
 
 def get_file_contents(filename, trim=False):
-    with open(os.path.join(KEYS_FOLDER, filename), "r") as f:
+    with open(os.path.join(KEYS_FOLDER, filename), "r", encoding="utf-8") as f:
         data = f.read()
         if trim:
             data = data.rstrip("\r\n")
@@ -41,6 +41,7 @@ class IntegrationTestCase(unittest.TestCase):  # pylint: disable=too-many-public
         self.last_response = None
         self.last_csrf_token = None
         self.redirect_url = None
+        self.last_response_headers = None
 
         # Perform setup steps
         self._set_up_app()
