+++ tests/integration/session/test_multiple_login.py
@@ -17,6 +17,7 @@ class MultipleClientTestCase(IntegrationTestCase):
         token = self.token_generator.create_token(schema_name, **payload_kwargs)
         self.get(client, "/session?token=" + token)
 
+    # pylint: disable=arguments-renamed
     def get(self, client, url, **kwargs):
         response = client.get(url, follow_redirects=True, **kwargs)
